requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
fake-useragent==1.4.0
pandas==2.1.3
python-dotenv==1.0.0
webdriver-manager==4.0.1
lxml==4.9.3
colorama==0.4.6
tqdm==4.66.1
schedule==1.2.0
pymongo==4.6.1
webdriverdownloader==1.1.0.3
flask==2.0.1
werkzeug==2.0.2
gunicorn==20.1.0

# Added for better anti-bot measures
undetected-chromedriver==3.5.4
selenium-stealth==1.0.6
random-user-agent==1.0.1

# For site verification and rate limiting
certifi==2023.11.17

# For improved output formatting
tabulate==0.9.0

# For concurrent processing
concurrent-log-handler==0.9.24

# For proxy management and CAPTCHA avoidance
PySocks==1.7.1
requests[socks]==2.31.0
python-anticaptcha==1.0.0

# For API integrations
aiohttp==3.8.5
tenacity==8.2.3
validators==0.22.0

# For email notifications
twilio==8.10.0
sendgrid==6.10.0

# For data analysis and visualization
matplotlib==3.8.0
numpy==1.26.2
Pillow==10.1.0

# For web interaction
requests-html==0.10.0

# For database operations
SQLAlchemy==2.0.23
